requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
websockets==13.1
chromadb==0.5.15
sentence-transformers==3.1.1
pyyaml==6.0.2
python-dotenv==1.0.1
python-telegram-bot==21.6
